+++ src/UnitTestEx/AspNetCore/HttpTesterBaseT2.cs
@@ -28,7 +28,7 @@ namespace UnitTestEx.AspNetCore
         /// <param name="testServer">The <see cref="TestServer"/>.</param>
         internal HttpTesterBase(TesterBase owner, TestServer testServer) : base(owner, testServer)
         {
-            _exceptionSuccessExpectations = new ExceptionSuccessExpectations(Owner);
+            _exceptionSuccessExpectations = new ExceptionSuccessExpectations(Implementor);
             _httpResponseExpectations = new HttpResponseExpectations(Owner);
             _responseValueExpectations = new ResponseValueExpectations<TSelf, TValue>(Owner, (TSelf)this);
             _eventExpectations = new EventExpectations(Owner);
@@ -55,6 +55,23 @@ namespace UnitTestEx.AspNetCore
         {
             UserName = userName;
             return (TSelf)this;
+        }
+
+        /// <summary>
+        /// Sets (overrides) the test user name (defaults to <see cref="TesterBase.UserName"/>).
+        /// </summary>
+        /// <param name="userIdentifier">The test user identifier.</param>
+        /// <returns>The <typeparamref name="TSelf"/> instance to support fluent-style method-chaining.</returns>
+        /// <remarks>The <see cref="TestSetUp.UserNameConverter"/> is required for the conversion to take place.</remarks>
+        public TSelf WithUser(object? userIdentifier)
+        {
+            if (userIdentifier == null)
+                return WithUser(null);
+
+            if (Owner.SetUp.UserNameConverter == null)
+                throw new InvalidOperationException($"The {nameof(TestSetUp)}.{nameof(TestSetUp.UserNameConverter)} must be defined to support user identifier conversion.");
+
+            return WithUser(Owner.SetUp.UserNameConverter(userIdentifier));
         }
 
         /// <inheritdoc/>